provision/.requirements.txt
ansible==2.9.12
asteval==0.9.18
certifi==2020.6.20
chardet==3.0.4
docker==4.3.0
dynaconf==3.1.0
idna==2.10
psutil==5.7.2
python-box==5.1.1
pywin32==227; sys_platform == "win32"
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0; platform_python_implementation == "CPython" and python_version < "3.9"
six==1.15.0
taskipy==1.3.0
toml==0.10.1
tomlkit==0.7.0
urllib3==1.25.10
websocket-client==0.57.0
